# GO-GRPC-API-GATEWAY/cmd/docs/swagger.yaml
definitions:
...
  models.ChatRequest:
    properties:
      friendID:
        type: string
      limit:
        type: string
      offset:
        type: string
    required:
    - friendID
    - limit
    - offset
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /chat/message:
    get:
      consumes:
      - application/json
      description: Retrieve UsersChats
      parameters:
      - description: GetChat details
        in: body
        name: chatRequest
        required: true
        schema:
          $ref: '#/definitions/models.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Users Chats
      tags:
      - Chat
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"